backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.1

# AI/LLM
anthropic==0.42.0
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-chroma==0.2.0
chromadb==0.5.23

# ML/DL
numpy==2.2.1
pandas==2.2.3
scikit-learn==1.6.0

# Image Processing
Pillow==11.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
sse-starlette==2.2.1
